Add Docker label for product-configure schema versions
The built image has a
`za.ac.kat.sdp.katsdpcontroller.product-configure-versions` label which
contains a space-separated list of schema versions supported. I chose
space- rather than comma-separation for consistency with
`za.ac.kat.sdp.image-depends`. This isn't yet used for anything, but
could be queried by the master controller to report to a client which
versions are available.

Relates to NGC-906.
